refactor(app): group by packages
