+++ app/src/main/java/fr/carbonit/th/Application.java
@@ -1,6 +1,8 @@
 package fr.carbonit.th;
 
 import fr.carbonit.th.configuration.TreasureHuntConfiguration;
+import fr.carbonit.th.exceptions.NoParameterException;
+import fr.carbonit.th.exceptions.TooManyParametersException;
 import fr.carbonit.th.parser.ConfigurationParser;
 import fr.carbonit.th.reader.ConfigurationReader;
