+++ app/src/main/java/fr/carbonit/th/simulation/models/Adventurer.java
@@ -1,6 +1,8 @@
-package fr.carbonit.th.simulation;
+package fr.carbonit.th.simulation.models;
 
 import fr.carbonit.th.simulation.actions.Action;
+import fr.carbonit.th.simulation.localisation.Coordinates;
+import fr.carbonit.th.simulation.localisation.Position;
 
 import java.util.Queue;
 import java.util.UUID;
