+++ app/src/main/java/fr/carbonit/th/simulation/models/Mountain.java
@@ -1,4 +1,6 @@
-package fr.carbonit.th.simulation;
+package fr.carbonit.th.simulation.models;
+
+import fr.carbonit.th.simulation.localisation.Coordinates;
 
 import java.util.Objects;
 import java.util.UUID;
