+++ app/src/test/java/fr/carbonit/th/ApplicationTest.java
@@ -1,5 +1,7 @@
 package fr.carbonit.th;
 
+import fr.carbonit.th.exceptions.NoParameterException;
+import fr.carbonit.th.exceptions.TooManyParametersException;
 import fr.carbonit.th.parser.UnparsableFileException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
