+++ app/src/test/java/fr/carbonit/th/simulation/OrientationTest.java
@@ -1,12 +1,13 @@
 package fr.carbonit.th.simulation;
 
+import fr.carbonit.th.simulation.localisation.Orientation;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-import static fr.carbonit.th.simulation.Orientation.*;
+import static fr.carbonit.th.simulation.localisation.Orientation.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
