+++ app/src/test/java/fr/carbonit/th/simulation/PositionTest.java
@@ -1,5 +1,8 @@
 package fr.carbonit.th.simulation;
 
+import fr.carbonit.th.simulation.localisation.Coordinates;
+import fr.carbonit.th.simulation.localisation.Orientation;
+import fr.carbonit.th.simulation.localisation.Position;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -7,7 +10,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-import static fr.carbonit.th.simulation.Orientation.*;
+import static fr.carbonit.th.simulation.localisation.Orientation.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PositionTest {
